+++ app/src/main/java/space/kuz/calculator/MainActivity.java
@@ -13,30 +13,25 @@ import android.widget.Toast;
 
 import java.util.ArrayList;
 
+import static space.kuz.calculator.CalculatingEquality.EMPTY;
+
 public class MainActivity extends AppCompatActivity {
-    ArrayList<String> textSing;     // Основная фармула
-    ArrayList<String> subTextSing; // Часть основной формулы (что находться в скобочках)
     private EditText basicEditText;
-    private Float result; // хранитель результата
     private boolean checkPoint = true;  // флаг для правильного набора дробных чисел
-    private static final String PLUS = "+";
-    private static final String MINUS = "-";
-    private static final String MULTIPLY = "*";
-    private static final String DIV = "÷";
-    private static final String POINT = ".";
-    private static final String LEFTSIGN = "(";
-    private static final String RIGHTSING = ")";
-    private static final String EMPTY = "";
+    CalculatingEquality calculatingEquality;
+    public static final String POINT = ".";
+
     
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        calculatingEquality = new CalculatingEquality();
         initButtonOnClickListener();
     }
     // Иницилизация всех кнопок и обработка их
     private void initButtonOnClickListener() {
-        basicEditText = (EditText) findViewById(R.id.basic_edit_text);
+        basicEditText = (EditText)findViewById(R.id.basic_edit_text);
         basicEditText.setText(EMPTY);
         basicEditText.setMovementMethod( new ScrollingMovementMethod());
         Button nullButton = findViewById(R.id.null_button);
@@ -77,8 +72,8 @@ public class MainActivity extends AppCompatActivity {
 
         pointButton.setOnClickListener(v -> {
             if (basicEditText.length() != 0 && checkPoint && checkSing()
-                    && !basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length()).equals(LEFTSIGN)
-                    && !basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length()).equals(RIGHTSING)
+                    && !isolationChar().equals(CalculatingEquality.LEFTSIGN)
+                    && !isolationChar().equals(CalculatingEquality.RIGHTSING)
             ) {
                 inputNumber(pointButton);
                 checkPoint = false;
@@ -118,10 +113,10 @@ public class MainActivity extends AppCompatActivity {
         });
 
         equalsButton.setOnClickListener(v -> {
-            convertFormula();
+           calculatingEquality.convertFormula(basicEditText);
             try {
-                calculateFormulaHard();
-            } catch (Exception e) {
+                calculatingEquality.calculateFormulaHard(basicEditText);
+           } catch (Exception e) {
                 Toast.makeText(this, "ОШИБКА!!! Введите заново", Toast.LENGTH_LONG).show();
             }
             checkPoint = true;
@@ -131,8 +126,8 @@ public class MainActivity extends AppCompatActivity {
             if (basicEditText.length() == 0) {
                 inputNumber(leftBracketButton);
             } else if (
-                    !isNumber(basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length())) &&
-                            !basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length()).equals(POINT)) {
+                    isNumber(isolationChar()) &&
+                            !isolationChar().equals(POINT)) {
                 inputNumber(leftBracketButton);
             }
             checkPoint = true;
@@ -141,8 +136,8 @@ public class MainActivity extends AppCompatActivity {
             if (basicEditText.length() == 0) {
                 inputNumber(rightBracketButton);
             } else if (
-                    !isNumber(basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length())) &&
-                            !basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length()).equals(POINT)) {
+                    isNumber(isolationChar()) &&
+                            !isolationChar().equals(POINT)) {
                 inputNumber(rightBracketButton);
             }
             checkPoint = true;
@@ -166,109 +161,7 @@ public class MainActivity extends AppCompatActivity {
     }
     // Метод проверки находиться рядом + - / * .
     private boolean checkSing() {
-        String endChar = basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length());
-        return !endChar.equals(PLUS) && !endChar.equals(MINUS) && !endChar.equals(MULTIPLY) && !endChar.equals(DIV) && !endChar.equals(POINT);
-    }
-    // Первод строки в колекцию для удобного вычисления
-    private void convertFormula() {
-        boolean checkNumber = false;
-        String Number = "";
-        StringBuilder formulaText = new StringBuilder(basicEditText.getText().toString());
-        textSing = new ArrayList();
-        String oneChar;
-        while (formulaText.length() != 0) {
-            oneChar = formulaText.substring(0, 1);
-            formulaText.delete(0, 1);
-            if (oneChar.equals(PLUS) || oneChar.equals(MINUS) || oneChar.equals(MULTIPLY) || oneChar.equals(DIV) || oneChar.equals(LEFTSIGN) || oneChar.equals(RIGHTSING)) {
-                checkNumber = true;
-                if (checkNumber && !Number.equals(EMPTY)) {
-                    textSing.add(Number);
-                    checkNumber = false;
-                    Number = EMPTY;
-                }
-                textSing.add(oneChar);
-
-            } else {
-                Number = Number + oneChar;
-                if (formulaText.length() == 0) {
-                    textSing.add(Number);
-                }
-            }
-        }
-    }
-    // Метод вычесления сложных формул со скобками
-    private void calculateFormulaHard() {
-        while (textSing.size() != 1) {
-            if (textSing.indexOf(RIGHTSING) == -1 && textSing.indexOf(LEFTSIGN) == -1) {
-                calculateFormulaSimple(textSing);
-            } else {
-
-                subTextSing = new ArrayList<>();
-                int endLeftSing = textSing.lastIndexOf(LEFTSIGN);
-                int k = endLeftSing + 1;
-                while (!textSing.get(k).equals(RIGHTSING)) {
-                    subTextSing.add(textSing.get(k));
-                    k++;
-                }
-                while (!textSing.get(endLeftSing).equals(RIGHTSING)) {
-                    textSing.remove(endLeftSing);
-                }
-                textSing.set(endLeftSing, calculateFormulaSimple(subTextSing));
-
-            }
-        }
-        writeEditText();
-    }
-    // Метод для вычесления простых формул (внутри скобки без скобок)
-    private String calculateFormulaSimple(ArrayList<String> subTextSing) {
-        // Условие для проверки если "-" в самом начале
-        checkFirstMinus(subTextSing);
-        // Операции в формуле
-        oneOperation(MULTIPLY, subTextSing);
-        oneOperation(DIV, subTextSing);
-        oneOperation(MINUS, subTextSing);
-        oneOperation(PLUS, subTextSing);
-        // Вывод результата
-        return subTextSing.get(0);
-    }
-    // Вычесляет элементарную операцию -*+/
-    private void oneOperation(String sing, ArrayList<String> subTextSing) {
-        int indexCalculate = subTextSing.indexOf(sing);
-        while (indexCalculate != -1 && subTextSing.size() > 1) {
-            switch (sing) {
-                case (MULTIPLY):
-                    result = Float.valueOf(subTextSing.get(indexCalculate - 1)) * Float.valueOf(subTextSing.get(indexCalculate + 1));
-                    break;
-                case (DIV):
-                    result = Float.valueOf(subTextSing.get(indexCalculate - 1)) / Float.valueOf(subTextSing.get(indexCalculate + 1));
-                    break;
-                case (MINUS):
-                    result = Float.valueOf(subTextSing.get(indexCalculate - 1)) - Float.valueOf(subTextSing.get(indexCalculate + 1));
-                    break;
-                case (PLUS):
-                    result = Float.valueOf(subTextSing.get(indexCalculate - 1)) + Float.valueOf(subTextSing.get(indexCalculate + 1));
-                    break;
-            }
-            subTextSing.set(indexCalculate, result + EMPTY);
-            subTextSing.remove(indexCalculate + 1);
-            subTextSing.remove(indexCalculate - 1);
-            indexCalculate = subTextSing.indexOf(sing);
-        }
-    }
-    // Запить в EditText
-    private void writeEditText() {
-        basicEditText.setText(EMPTY);
-        for (String s : textSing
-        ) {
-            basicEditText.setText(basicEditText.getText() + s);
-        }
-    }
-    // Проверка наличия первого минуса в строке
-    private void checkFirstMinus(ArrayList<String> subTextSing) {
-        if (subTextSing.get(0).equals(MINUS)) {
-            subTextSing.set(1, String.valueOf(Float.valueOf(subTextSing.get(1)) * -1));
-            subTextSing.remove(0);
-        }
+        return !isolationChar().equals(CalculatingEquality.PLUS) && !isolationChar().equals(CalculatingEquality.MINUS) && !isolationChar().equals(CalculatingEquality.MULTIPLY) && !isolationChar().equals(CalculatingEquality.DIV) && !isolationChar().equals(POINT);
     }
     // Проверка является строка числом
     private static boolean isNumber(String string) {
@@ -279,6 +172,10 @@ public class MainActivity extends AppCompatActivity {
             return false;
         }
     }
-    // Метод уменьшения шрифта
+
+    private String isolationChar(){
+        return basicEditText.getText().toString().substring(basicEditText.getText().length() - 1, basicEditText.getText().length());
+    }
+
 
 }
